cleanup and rootless test coverage
